requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
spacy>=3.7.0
nltk>=3.8.0
textstat>=0.7.0
chardet>=5.0.0
numpy>=1.24.0
boto3>=1.34.0
google-cloud-vision>=3.4.0
google-oauth2-tool>=0.0.3
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
requests>=2.31.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0 